2019 Code generating a BTC public and private key
Code from 2019 that has not been updated or ran since; the code was mostly in Swift, with small bits of C# to configure some of the button logic for iOS. 

The project was developed in MacOS's Xcode IDE, and was not updated with any styling, but basic functionality. Has to open the project from the workspace file directly to get it to run and compile properly. May had something to do with the pods, but stopped there.